refactor class names
